+++ backend/canisters/group_index/impl/src/updates/c2c_notify_low_balance.rs
@@ -48,7 +48,7 @@ fn prepare(runtime_state: &RuntimeState) -> Result<PrepareResult, NotifyLowBalan
 }
 
 fn commit(canister_id: CanisterId, top_up: CyclesTopUp, runtime_state: &mut RuntimeState) {
-    runtime_state.data.total_cycles_spent_on_canisters += top_up.amount as u128;
+    runtime_state.data.total_cycles_spent_on_canisters += top_up.amount;
     runtime_state
         .data
         .local_index_map
